GEODE-10166: Move tests to correct directory
 - Rename InternalCacheForClientAccessDUnitTest ->
 InternalCacheForClientAccessDistributedTest
 - Move InternalCacheForClientAccessDistributedTest to geode-core module
 - Remove spurious <tearDown> parameterization from
 InternalCacheForClientAccessDistributedTest
 - Rename NestedQueryClassCastExceptionFailureDUnitTest ->
 NestedQueryClassCastExceptionFailureDistributedTest
 - Move NestedQueryClassCastExceptionFailureDistributedTest to
   geode-core module, change package to cache.query.dunit
 - Rename RebalanceCommandDistributesTest to
  RelanaceCommandAcceptanceTest
 - Move RebalanceCommandAcceptanceTest to geode-assembly module
  acceptanceTest directory as it requires that GEODE_HOME is set
 - Remove files from sanctioned-geode-dunit-serializables.txt

Authored-by: Donal Evans <doevans@vmware.com>
